+++ src/Console/CreateBlockCommand.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Thinhnx\LaravelPageBuilder\Console;
 
 use Illuminate\Console\Command;
+use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\File;
 use Throwable;
@@ -42,6 +43,7 @@ class CreateBlockCommand extends Command
         DB::transaction(function () use ($data) {
             app(config('page-builder.models.block'))::create($data);
             File::put(resource_path('views/vendor/page-builder/blocks/' . $data['type'] . '.blade.php'), '');
+            Cache::forget(config('page-builder.cache.keys.blocks'));
         });
         $this->info('Created the block successfully!!!');
     }